Fixed Everything
Added Two Tests. Fixed the variable names in Parser methods, but unsure if that will ultimately be how we do that. Fixed some issues in Main so it would not be so angry all the time. Cleaned up POM.xml and fixed misc so that Payton will stop having language issues. Fixed Webpage and added  semi-pretty method to print itself.